Stoichiometric coefficients

Coefficients for calculating substrate oxidation from gas exchange, as derived and presented in Telmosse (2022).

Usage

stoich_coefficients

Format

A named list with coefficients for CHO and fat calculations

References

Telmosse, E. (2022). Contribution des glucides exogènes à la fourniture d’énergie lors d’un effort prolongé : analyse par régression multiple. Mémoire de maîtrise, Université de Montréal. Livesey, G., & Elia, M. (1988). Estimation of energy expenditure, net carbohydrate utilization, and net fat oxidation and synthesis by indirect calorimetry: evaluation of errors with special reference to the detailed composition of fuels. The American Journal of Clinical Nutrition, 47(4), 608–628. Institute of Medicine, Food and Nutrition Board. (2005). Dietary Reference Intakes for Energy, Carbohydrate, Fiber, Fat, Fatty Acids, Cholesterol, Protein, and Amino Acids. National Academies Press. Péronnet, F., & Massicotte, D. (1991). Table of nonprotein respiratory quotient: an update. Canadian Journal of Sport Sciences = Journal Canadien Des Sciences Du Sport, 16(1), 23–29.